requirements.txt
Appium-Python-Client==4.4.0
selenium==4.27.1
colorlog-=6.9.0